Api call added

See merge request newslatelabs/insite-mobile!200

+++ lib/core/models/estimated_cycle_volume_payload.dart
@@ -37,13 +37,13 @@ class AssetProductivitySettings {
 
 @JsonSerializable()
 class Cycles {
-  int sunday;
-  int monday;
-  int tuesday;
-  int wednesday;
-  int thursday;
-  int friday;
-  int saturday;
+  double sunday;
+  double monday;
+  double tuesday;
+  double wednesday;
+  double thursday;
+  double friday;
+  double saturday;
 
   Cycles(
       {this.sunday,
